refactor: 카드 뽑기 테스트를 Player 테스트로 이동
