+++ src/test/java/blackjack/domain/player/ChallengerTest.java
@@ -18,29 +18,6 @@ import org.junit.jupiter.params.provider.MethodSource;
 class ChallengerTest {
 
     @Test
-    @DisplayName("플레이어는 초기 카드 2장을 받는다")
-    void get_two_cards() {
-        Player player = new Challenger("oing");
-        Card card1 = new Card(Shape.HEART, Number.FOUR);
-        Card card2 = new Card(Shape.CLOVER, Number.KING);
-        player.pickStartCards(List.of(card1, card2));
-
-        assertThat(player.getHoldingCards().getCards())
-                .containsExactly(card1, card2);
-    }
-
-    @Test
-    @DisplayName("추가 카드를 뽑는다.")
-    void pick_card() {
-        Player player = new Challenger("ditoo");
-        Card card = new Card(Shape.DIAMOND, Number.JACK);
-        player.pick(card);
-
-        assertThat(player.getHoldingCards().getCards())
-                .contains(card);
-    }
-
-    @Test
     @DisplayName("이름이 '딜러'인 경우 예외가 발생한다")
     void validate_name() {
         assertThrows(InvalidPlayerNameException.class,
